refactor(public): split event detail page under the line cap (#544)
Extract the public event-detail page (1025 lines) into two co-located
cohesive units under $lib/components/events/, behavior-preserving:

- event-checkout-controller.svelte.ts — rune-based factory owning the
  ticket claim/checkout/PWYC/resume/cancel mutations and their success/
  error side effects (createCheckoutController); host wires narrow
  getters/setters. #535 error-detail helper moves intact.
- EventConfirmationBanners.svelte — the four post-redirect alert banners
  (payment success/cancelled, RSVP, ticket) behind boolean props.

Template sections here already delegate to separate components; the
script (checkout logic, 617 lines) was the bulk, so the controller
factory (method c) is the load-bearing seam. SSR/SEO untouched: SeoHead
placement byte-identical, no ssr override, no inline JSON-LD moved.
Event detail page 1025 -> 637 lines.
